Add support for adapters with multiple frontends
This is based on a patch by Simon Liddicott which I got from
http://gt.owl.de/pipermail/getstream/2010-February/000138.html

In contrast to that patch, I avoided the parsing of a new config value
type "LCV_ADAPTER". Instead, I added a new numeric parameter "frontend"
in the config group "adapter".

This, of course, means that the same adapter id now can occur more than
once, so the config parser can't detect bad configurations containing
duplicate adapters any more. To me this seems to be a fair tradeoff to
avoid the parsing of adapter:frontend pairs.